requirements.txt
flask
requests
streamlit
google-generativeai
watchdog
